+++ core/interfaces/src/application.rs
@@ -162,9 +162,9 @@ pub trait SyncQueryRunnerInterface: Clone + Send + Sync + 'static {
     /// returns the service information for a given serviceid
     fn get_service_info(&self, service_id: ServiceId) -> Service;
 
-    fn pubkey_to_index(&self, node: NodePublicKey) -> Option<u64>;
+    fn pubkey_to_index(&self, node: NodePublicKey) -> Option<u32>;
 
-    fn index_to_pubkey(&self, node_index: u64) -> Option<NodePublicKey>;
+    fn index_to_pubkey(&self, node_index: u32) -> Option<NodePublicKey>;
 }
 
 #[derive(Clone, Debug)]
